no exit

+++ app/src/main/java/com/cod5/signyourapk/MainActivity.java
@@ -1,15 +1,12 @@
+/*
+ * 25 June MMXXIV PUBLIC DOMAIN by JML
+ *
+ * The authors disclaim copyright to this source code
+ *
+ */
 package com.cod5.signyourapk;
 
-import androidx.activity.result.ActivityResult;
-import androidx.activity.result.ActivityResultLauncher;
-import androidx.annotation.RequiresApi;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.app.ActivityCompat;
-
-import android.app.Activity;
-import android.content.Context;
 import android.content.Intent;
-import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
 import android.content.res.AssetManager;
 import android.net.Uri;
@@ -17,29 +14,29 @@ import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
 import android.provider.Settings;
-import android.util.Log;
 import android.view.View;
 import android.widget.RadioButton;
-import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.RequiresApi;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.ActivityCompat;
 
 import com.android.apksigner.ApkSignerTool;
 import com.cod5.signyourapk.databinding.ActivityMainBinding;
-
-import org.conscrypt.BuildConfig;
 
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
-import java.util.Objects;
+import java.nio.file.Files;
 
 public class MainActivity extends AppCompatActivity {
 
     private static final int STORAGE_PERMISSION_CODE = 189;
 
-    // Used to load the 'signyourapk' library on application startup.
     static {
         System.loadLibrary("signyourapk");
     }
@@ -53,16 +50,11 @@ public class MainActivity extends AppCompatActivity {
         binding = ActivityMainBinding.inflate(getLayoutInflater());
         setContentView(binding.getRoot());
 
-        /*TextView tv = binding.sampleText;
-        tv.setText(stringFromJNI());*/
-        binding.sampleText.setText("Sign you APK");
-        binding.button.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                onClickMe(v);
-            }
-        });
-        hasWriteStoragePermission();
+        binding.sampleText.setText(R.string.sign_you_apk);
+        binding.button.setOnClickListener(this::onClickMe);
+        if (hasWriteStoragePermission()) {
+            binding.sampleText.setText(R.string.sign_you_apk);
+        }
         askAllFilesPermission();
         listDownloadsFiles();
     }
@@ -72,21 +64,23 @@ public class MainActivity extends AppCompatActivity {
         File d = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS);
         File[] lst = d.listFiles();
         int i  = 0;
-        for(File f: lst) {
-            if (f.getName().endsWith(".apk")) {
-                RadioButton r;
-                r = new RadioButton(this);
-                r.setText(f.getAbsolutePath());
-                binding.radio.addView(r, i);
-                i++;
+        if (lst != null) {
+            for(File f: lst) {
+                if (f.getName().endsWith(".apk")) {
+                    RadioButton r;
+                    r = new RadioButton(this);
+                    r.setText(f.getAbsolutePath());
+                    binding.radio.addView(r, i);
+                    i++;
+                }
             }
         }
 
     }
     /* print result of permission request */
     public void onRequestPermissionsResult(int requestCode,
-                                            String[] permissions,
-                                            int [] grantResults) {
+                                           @NonNull String[] permissions,
+                                           @NonNull int [] grantResults) {
         super.onRequestPermissionsResult(requestCode, permissions, grantResults);
         if (requestCode == STORAGE_PERMISSION_CODE) {
             if (grantResults.length > 0 && grantResults[0] == PackageManager.PERMISSION_GRANTED) {
@@ -96,10 +90,10 @@ public class MainActivity extends AppCompatActivity {
             }
         }
     }
-    private boolean askAllFilesPermission() {
+    private void askAllFilesPermission() {
         if (Build.VERSION.SDK_INT >= 30) {
             if (hasAllFilesPermission()) {
-                return true;
+                return;
             }
             startActivity(
                     new Intent(
@@ -108,7 +102,6 @@ public class MainActivity extends AppCompatActivity {
                     )
             );
         }
-        return true;
     }
 
 
@@ -145,31 +138,46 @@ public class MainActivity extends AppCompatActivity {
             Toast.makeText(MainActivity.this, "please restart app.", Toast.LENGTH_LONG).show();
             return;
         } else {
-            binding.sampleText.setText("Start...");
+            binding.sampleText.setText(R.string.start);
             Toast.makeText(MainActivity.this, "Start signing", Toast.LENGTH_LONG).show();
         }
         try {
-            InputStream in = null;
-            OutputStream out = null;
+            InputStream in;
+            OutputStream out;
             AssetManager assetManager = getAssets();
-            in = assetManager.open("cert.bks");
-            out = new FileOutputStream(this.getApplicationInfo().dataDir.toString() + "/cert.bks");
-            copyFile(in, out);
+            File cert = new File(Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS)+"/cert.bks");
+            if (!cert.exists()) {
+                in = assetManager.open("cert.bks");
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+                    out = Files.newOutputStream(cert.toPath());
+                } else {
+                    out = new FileOutputStream(cert);
+                }
+                copyFile(in, out);
+                out.close();
+            }
             int id = binding.radio.getCheckedRadioButtonId();
             if (id >= 0) {
                 RadioButton rdb = binding.radio.findViewById(id);
                 ApkSignerTool.main(new String[]{"sign", "--min-sdk-version", "16", "--ks",
-                        this.getApplicationInfo().dataDir.toString() + "/cert.bks",
-                        "--ks-pass", "pass:12345678", rdb.getText().toString()});
-                binding.sampleText.setText("APK signed !");
+                        cert.getPath(),
+                        "--ks-pass", "pass:"+ binding.passwd.getText().toString(),
+                        rdb.getText().toString()});
+                binding.sampleText.setText(R.string.apk_signed);
                 Toast.makeText(MainActivity.this, "Success!", Toast.LENGTH_LONG).show();
+            } else {
+                binding.sampleText.setText(R.string.please_select_an_apk);
             }
         } catch (Exception e) {
             e.printStackTrace();
-            binding.sampleText.setText(e.getCause().toString());
+            if (e.getCause() != null) {
+                binding.sampleText.setText(e.getCause().toString());
+            } else {
+                binding.sampleText.setText(R.string.failed);
+            }
         }
     }
-    // Method used by copyAssets() on purpose to copy a file.
+
     private void copyFile(InputStream in, OutputStream out) throws IOException {
         byte[] buffer = new byte[1024];
         int read;
@@ -181,13 +189,9 @@ public class MainActivity extends AppCompatActivity {
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
         if(requestCode == 123 && resultCode == RESULT_OK) {
-            binding.sampleText.setText(this.getExternalFilesDir(Environment.DIRECTORY_DOWNLOADS)
-                    + "/app.apk");
+            binding.sampleText.setText(String.format("%s/app.apk", this.getExternalFilesDir(Environment.DIRECTORY_DOWNLOADS)));
         }
     }
-    /**
-     * A native method that is implemented by the 'signyourapk' native library,
-     * which is packaged with this application.
-     */
-    public native String stringFromJNI();
+
+    //public native String stringFromJNI();
 }
